sysrq: Skip synchronize_rcu() if there is no old op
synchronize_rcu() is expensive. Currently it is called as part of the sysrq
registration/unregistration, which happens during boot several times.
Now, the reason for the synchronize_rcu() is to allow an old registered
operation to expire properly... which is pointless if the old operation
is NULL...
So we can save the common case of the old operation being NULL a lot of time
by just checking for non-NULL prior to the synchronize_rcu()

Signed-off-by: Arjan van de Ven <arjan@linux.intel.com>
Signed-off-by: Miguel Bernal Marin <miguel.bernal.marin@linux.intel.com>
[nc: Applied from clearlinux-pkgs directly but this is in the stock WSL2
     kernel]
Signed-off-by: Nathan Chancellor <nathan@kernel.org>
Signed-off-by: Fajar Ibrahim <fajaribrahim0031@gmail.com>