f2fs/rapid_gc: do not wait until all invalid blocks are discarded
Undiscarded blocks may not reach 0 for various reasons, such as
discard granularity settings and discard thread terminating before GC ends,
and when that happens, GC thread enters an indefinite loop.

Remove this logic entirely.

This is not much of an issue since discard thread does not rely on rapid GC.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Change-Id: Ie2c32f2c3aca0684518404d30ff4698b832a71fa
Signed-off-by: fajar <santuyz321@gmail.com>