ARM64/configs: ginkgo: Make ion allocate in page-sized chunks
With this option disabled, Ion allocates memory in order-4 sized chunks
(16 pages at a time), which is quite a heavy burden on the page
allocator, and can take a while to allocate when memory is fragmented
after a lot of uptime. To lower the allocation latency in Ion, have it
only allocate in page-sized chunks.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[yaro: add kconfig entry too]
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: fajar <santuyz321@gmail.com>