ARM64/dts: qcom: Disable CoreSight DT entries for trinket
CoreSight is a set of trace and debug software by Arm Limited. and is used
on their architecture of SoCs and is very instrumental in early bring-up or
in development stage of boards. These tracing features are useful in such
scenarios but it comes with its heavy overhead.

CoreSight tracing configs are enabled by default on perf-defconfigs but is
recommended to be disabled on production devices as per 80-PJ035-1 article
due its heavy overhead, even when these debugging configs are disabled due
to these DT entries clocks are being left on which in turn introduces power
regressions.

This commit marks JTAG driver nodes as disabled and removes the inclusion of
the coresight devicetree with its properties and mappings, also remove other
nodes that were reduntant due to this change.

Signed-off-by: Daniel Jacob Chittoor <djchittoor47@gmail.com>
[ghostrider-reborn: adapt for trinket]
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>
Signed-off-by: fajar <santuyz321@gmail.com>